Backup example where VM gets shut down before backup
Here the VM first gets shut down. If it does shutdown it will make a snapshot and turn on the vm again so that the VM is not shut down for long. After this the VM get copied and the snapshot is committed. 
Then a the whole VM is compressed to a .tar.gz file and an md5sum is made of the compressed file.